Split captions into sentences

Split figure, table and supplementary material captions into sentences

pmc_caption(doc)

Arguments

doc
xml_document from PubMed Central

Value

a tibble with tag, label, sentence number and text

Author

Chris Stubben

Examples

# doc <- pmc_xml("PMC2231364") # OR
doc <- xml2::read_xml(system.file("extdata/PMC2231364.xml",
  package = "tidypmc"
))
x <- pmc_caption(doc)

#> Found 5 figures

#> Found 4 tables

#> Found 3 supplements

x

#> # A tibble: 30 x 4
#>    tag    label   sentence text                                                 
#>    <chr>  <chr>      <int> <chr>                                                
#>  1 figure Figure…        1 "Environmental modulation of expression of virulence…
#>  2 figure Figure…        2 "Shown in the squares are the putative stages of tra…
#>  3 figure Figure…        3 "The TreeView charts show the transcriptional change…
#>  4 figure Figure…        4 "Color intensities denote log2 ratios as follows: gr…
#>  5 figure Figure…        1 "RT-PCR analysis of potential operons."              
#>  6 figure Figure…        2 "Shown is the electrophoresis image of an RT-PCR pro…
#>  7 figure Figure…        3 "Total RNA was used to synthesize cDNA in the presen…
#>  8 figure Figure…        4 "Genomic DNA was used as a template, and is indicate…
#>  9 figure Figure…        5 "\"Marker\" represents a DNA size marker (900, 700, …
#> 10 figure Figure…        1 "Schematic representation of the clustered microarra…
#> # … with 20 more rows

dplyr::filter(x, sentence == 1)

#> # A tibble: 12 x 4
#>    tag      label            sentence text                                      
#>    <chr>    <chr>               <int> <chr>                                     
#>  1 figure   Figure 1                1 Environmental modulation of expression of…
#>  2 figure   Figure 2                1 RT-PCR analysis of potential operons.     
#>  3 figure   Figure 3                1 Schematic representation of the clustered…
#>  4 figure   Figure 4                1 Graphical representation of the consensus…
#>  5 figure   Figure 5                1 EMSA analysis of the binding of Fur prote…
#>  6 table    Table 1                 1 Stress-responsive operons in Y. pestis pr…
#>  7 table    Table 2                 1 Classification of the gene members of the…
#>  8 table    Table 3                 1 Motif discovery for the clustering genes  
#>  9 table    Table 4                 1 Designs for expression profiling of Y. pe…
#> 10 supplem… Additional file…        1 Growth curves of Y. pestis strain 201 und…
#> 11 supplem… Additional file…        1 All the transcriptional changes of 4005 g…
#> 12 supplem… Additional file…        1 List of oligonucleotide primers used in t…